decomp progress: 9 functions to 100% + many large partials
100%:
- Quazal::MemoryManager::BeginProtection 70.0% → 100% (mutex local hoist)
- Quazal::MemoryManager::EndProtection 70.0% → 100%
- Quazal::Platform::Sleep 67.8% → 100% (s64 cast)
- Quazal::Job::SetToComplete 67.6% → 100% (completion callback)
- Quazal::Job::SetToCancel 67.6% → 100%
- Cheats::ForceGPHangRecoveryDF 76.5% → 100% (split bool flags)
- SystemTerminate 74.9% → 100% (uncomment 14 missing Terminate calls)
- TrackPanel::CleanUpTracks 75.1% → 100% (TrimExcess vs ClearAndShrink)
- BandCamShot::OnTestDelta 71.8% → 100% (rewrite Min/MaxTime logic)
- GemTrackDir::UpdateFingerFeedback 93.9% → 100% (default-then-override)

Partial (10+ point gains):
- Gem::AddWidgetInstanceImpl 75.5% → 99.4%
- Gem::operator= 76.7% → 98.6% (correct copied fields)
- TourPerformerImpl::GetQuestSuccessfulSongCount 71.3% → 85.4%
- BandTrack::SyncInstrument 92.0% → 99.9% (MILO_NOTIFY_ONCE wrap)
- ScrollbarDisplay::UpdateThumbScaleAndPosition 85.3% → 97.0%
- Job::Resume/SetToReady/SetToRunning/SetToInitial/SetToSuspended 70.9% → 92.4%
- NetCacheMgr::CheatNextServer 71.6% → 89.2%

+++ src/network/Platform/Platform.cpp
@@ -65,7 +65,7 @@ namespace Quazal {
 
     void Platform::Sleep(uint msecs) {
         if (msecs != 0) {
-            OSSleepTicks(OSMillisecondsToTicks(msecs));
+            OSSleepTicks(OSMillisecondsToTicks((s64)msecs));
         } else
             OSYieldThread();
     }
